deployment/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
openai-whisper>=20231117
opencv-python>=4.8.0
numpy>=1.24.0
nvgpu==0.10.0
soundfile>=0.12.1
ffmpeg-python>=0.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6